refactor(configure): move the Session Manager plugin check out of init
The plugin is a per-machine tool for `run` / `db:tunnel` — same cadence
and kind as the aws/jq/op binaries configure already checks, not part of
app scaffolding. Left in init, a developer who joins an existing app
(clone + configure, never init) never got the check and hit a cryptic
SSM error on their first `yolo run`. Now it rides configure's machine
setup, so every machine that authenticates also gets the plugin.
