+++ docs/reference/commands.md
@@ -68,8 +68,7 @@ Interactive. Prompts for the app name, the environment to add (e.g. `production`
 - Writes a default `Dockerfile` and `.dockerignore` (asks before overwriting existing ones).
 - Creates a starter `.env.<environment>`.
 - Appends `.yolo`, `.env.<environment>` (plus `.env.staging`/`.env.production`), and the env-shared working copies (`.env.environment.*`, `yolo-environment-*.yml`) to `.gitignore`.
-- Offers to install the AWS Session Manager plugin (used by [`run`](#yolo-run) and [`db:tunnel`](#yolo-db-tunnel)).
-- Offers to run [`configure`](#yolo-configure) at the end — the natural next step on a fresh app. Decline if this machine already has the account's profile set up.
+- Offers to run [`configure`](#yolo-configure) at the end — the natural next step on a fresh app (it handles machine setup, including the Session Manager plugin). Decline if this machine already has the account's profile set up.
 
 This is the only command that runs without an existing manifest.
 
@@ -90,14 +89,15 @@ yolo configure <environment> [--driver=<driver>]
 Interactive; each step is checked and offered a fix rather than left to fail later:
 
 1. **Binaries** — verifies `aws` (plus `jq` and `op` for the 1Password driver) and prints the Homebrew install lines for anything missing.
-2. **Helper install** (1Password driver) — copies `yolo-credentials` from the composer package to `~/.local/bin`, so the profile survives checkout moves and `composer update` refreshes reach it on the next run.
-3. **Item verification** (1Password driver) — confirms the named item exists and carries `aws_access_key_id` / `aws_secret_access_key` before anything is written.
-4. **Profile write** — writes `credential_process` + the manifest's region as `[profile <name>]` in `~/.aws/config`, replacing an existing block only after confirmation. Leftover `sso_*` keys are called out by name — the CLI resolves SSO configuration ahead of `credential_process`, so remnants silently break the setup.
-5. **Shadow check** — a same-named section in `~/.aws/credentials` takes precedence over `credential_process`; `configure` detects it and offers to remove it.
-6. **Wire and verify** — sets `YOLO_<ENVIRONMENT>_AWS_PROFILE` in the app's local `.env`, then proves the chain with `aws sts get-caller-identity` and holds the resolved account against the manifest's `account-id`.
-7. **MFA gate** — a green verify can't show whether MFA was forwarded (the helper only warns on stderr), so `configure` checks explicitly: is an MFA device registered on the IAM user, and does the 1Password item carry a TOTP to forward? [Every tier requires MFA](#conventions), so a missing device or TOTP **fails the command** with the missing half named — the alternative is an opaque AccessDenied at the first real command.
+2. **Session Manager plugin** — a per-machine tool [`run`](#yolo-run) and [`db:tunnel`](#yolo-db-tunnel) need to reach a running container; offers to install it (non-fatal — it isn't needed for `configure` itself).
+3. **Helper install** (1Password driver) — copies `yolo-credentials` from the composer package to `~/.local/bin`, so the profile survives checkout moves and `composer update` refreshes reach it on the next run.
+4. **Item verification** (1Password driver) — confirms the named item exists and carries `aws_access_key_id` / `aws_secret_access_key` before anything is written.
+5. **Profile write** — writes `credential_process` + the manifest's region as `[profile <name>]` in `~/.aws/config`, replacing an existing block only after confirmation. Leftover `sso_*` keys are called out by name — the CLI resolves SSO configuration ahead of `credential_process`, so remnants silently break the setup.
+6. **Shadow check** — a same-named section in `~/.aws/credentials` takes precedence over `credential_process`; `configure` detects it and offers to remove it.
+7. **Wire and verify** — sets `YOLO_<ENVIRONMENT>_AWS_PROFILE` in the app's local `.env`, then proves the chain with `aws sts get-caller-identity` and holds the resolved account against the manifest's `account-id`.
+8. **MFA gate** — a green verify can't show whether MFA was forwarded (the helper only warns on stderr), so `configure` checks explicitly: is an MFA device registered on the IAM user, and does the 1Password item carry a TOTP to forward? [Every tier requires MFA](#conventions), so a missing device or TOTP **fails the command** with the missing half named — the alternative is an opaque AccessDenied at the first real command.
 
-Profiles map to AWS **accounts**, not apps — run `configure` once per account, then reuse the profile name in each sibling app's `.env` (step 6 is the only per-app part). See [Developer Credentials](/guide/credentials).
+Profiles map to AWS **accounts**, not apps — run `configure` once per account, then reuse the profile name in each sibling app's `.env` (the `.env` wiring in step 7 is the only per-app part). See [Developer Credentials](/guide/credentials).
 
 ---
